pyproject.toml
[project]
name = "fakechecker"
version = "0.1.0"
description = "123"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.13.0",
    "aiojobs>=1.3.0",
    "environs>=11.0.0",
    "json-repair>=0.30.0",
    "langfuse>=2.52.1",
    "orjson>=3.10.7",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.7.1"
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"
# Enable the "ISORT" rule to automatically sort imports
select = ["I001"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = true
mark-parentheses = true

[tool.ruff.lint.flake8-bugbear]
# Allow default arguments like, e.g., `data: List[str] = fastapi.Query(None)`.
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.mypy]
[[tool.mypy.overrides]]
module = ["googleapiclient.*", "langfuse.*"]
ignore_missing_imports = true